Initial serving of generated HTML files now working
We look for the file in the directory, and if it exists we embed it in
the template using {% ssi %} - the settings change is required for ssi.

We also make our own dictionaries to hold the tag data needed in the
template, but we make them outside of the view.

The view is now quite a bit smaller as it doesn't get bits and bobs from
the database.

Commit also include various bits of refactoring.